bugflow/requirements.txt
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
idna==3.3
markovify==0.9.4
mistletoe==0.8.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
Unidecode==1.3.4
urllib3==1.26.9
wrapt==1.14.0
